reorganization of functions in modules
